feat(theme): add context
Create a React Context to handle theme changes.
