+++ src/shared/theme/types.ts
@@ -1,0 +1,2 @@
+export type Theme = 'dark' | 'light';
+export type ThemePreference = 'system' | Theme;
